requirements.txt
pyaudio
numpy
matplotlib
openai-whisper
torch
tkinter